requirements.txt
aiogram
python-dotenv
aiosqlite
aiomysql
sqlalchemy
aiohttp
cachetools
